requirements.txt
streamlit
google-generativeai
langfuse
feedparser
beautifulsoup4
reportlab
python-dotenv
pandas
plotly
PyPDF2
